@@ -20,7 +20,6 @@ namespace HappyCode.NetCoreBoilerplate.Api.IntegrationTests.Infrastructure
         public override void ConfigureServices(IServiceCollection services)
         {
             services.AddMvcCore()
-                .AddJsonFormatters()
                 .AddDataAnnotations()
                 .SetCompatibilityVersion(CompatibilityVersion.Latest);
 
@@ -41,7 +40,7 @@ namespace HappyCode.NetCoreBoilerplate.Api.IntegrationTests.Infrastructure
             // builder.RegisterType<SomeService>().As<ISomeService>();  //if needed override registration with own test fakes
         }
 
-        public override void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public override void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             var employeesContext = app.ApplicationServices.GetService<EmployeesContext>();
             EmployeesContextDataFeeder.Feed(employeesContext);
@@ -49,7 +48,11 @@ namespace HappyCode.NetCoreBoilerplate.Api.IntegrationTests.Infrastructure
             var carsContext = app.ApplicationServices.GetService<CarsContext>();
             CarsContextDataFeeder.Feed(carsContext);
 
-            app.UseMvc();
+            app.UseRouting();
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
         }
     }
 }
