@@ -1,9 +1,8 @@
-using System;
 using System.Net.Http;
 using Autofac.Extensions.DependencyInjection;
-using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.TestHost;
+using Microsoft.Extensions.Hosting;
 using Xunit;
 
 namespace HappyCode.NetCoreBoilerplate.Api.IntegrationTests.Infrastructure
@@ -14,15 +13,18 @@ namespace HappyCode.NetCoreBoilerplate.Api.IntegrationTests.Infrastructure
 
         public TestServerClientFixture()
         {
-            var server = new TestServer(WebHost.CreateDefaultBuilder()
-                .UseEnvironment("Test")
-                .ConfigureServices(services => services.AddAutofac())
-                .UseStartup<TestStartup>())
+            var host = new HostBuilder()
+                .UseServiceProviderFactory(new AutofacServiceProviderFactory())
+                .ConfigureWebHost(webBuilder =>
                 {
-                    BaseAddress = new Uri("http://localhost:5000")
-                };
+                    webBuilder
+                        .UseEnvironment("Test")
+                        .UseStartup<TestStartup>()
+                        .UseTestServer();
+                })
+                .Start();
 
-            Client = server.CreateClient();
+            Client = host.GetTestClient();
         }
     }
 
